Support FIPS mode operation in pgcrypto
In Greenplum 4.3 there was a GUC in pgcrypto controlling FIPS mode
operation which was deprecated. This brings the GUC back, as there
has been requests from production instances.

To enable the GUC, Greenplum must be dynamically linked against an
OpenSSL installation which is compiled with the FIPS module (1.0.2
with FIPS 2.0 is tested and supported, no FIPS module exist for
later versions of OpenSSL as of yet). There is an additional strict
mode as well which requires both OpenSSL to be FIPS capable as well
as the operating system kernel to be running in FIPS mode (this is
only supported on Linux for now).

Some ciphers will be disallowed in this mode even though they are
defined as legal by FIPS-140 level 2, due to the fact that pgcrypto
is using the lowlevel API for these and this is not certified for
FIPS operation.